+++ nanosvgrast.c
@@ -1,0 +1,6 @@
+#include <stdlib.h>
+#include <string.h>
+
+#include <nanosvg.h>
+#define NANOSVGRAST_IMPLEMENTATION
+#include <nanosvgrast.h>